MoleculeViewer/requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
rdkit==2024.9.1
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
